// src/offer-channels/offer-channel.enum.ts
export enum OfferChannelCode {
    IN_STORE = 'IN_STORE',
    WEBSITE = 'WEBSITE',
    APP_WOLT = 'APP_WOLT',
    APP_KASPI = 'APP_KASPI',
    MESSENGER_WHATSAPP = 'MESSENGER_WHATSAPP',
    MESSENGER_TELEGRAM = 'MESSENGER_TELEGRAM',
    MARKETPLACE = 'MARKETPLACE',
    PHONE = 'PHONE',
    SOCIAL_INSTAGRAM = 'SOCIAL_INSTAGRAM',
    SOCIAL_TIKTOK = 'SOCIAL_TIKTOK',
}